change type of identifier() to uintptr_t using uintptr_t in ~Collector class change Py_hash_t hash(PyObject* pyObj) function

+++ tests/libsample/collector.cpp
@@ -27,7 +27,7 @@ void Collector::clear()
     m_items.clear();
 }
 
-Collector& Collector::operator<<(unsigned long item)
+Collector& Collector::operator<<(uintptr_t item)
 {
     m_items.push_back(item);
     return *this;
@@ -39,7 +39,7 @@ Collector& Collector::operator<<(const ObjectType *obj)
     return *this;
 }
 
-std::list<unsigned long> Collector::items()
+std::list<uintptr_t> Collector::items()
 {
     return m_items;
 }
